requirements.txt
torch
torchvision
opencv-python
numpy
pillow
streamlit
scikit-learn
